GCP/requirements.txt
# Function dependencies, for example:
# package>=version

tensorflow==2.0.0b0
google-cloud-storage==1.25.0
Pillow==6.0.0
numpy==1.18.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
scipy==1.4.1
nltk==3.4.5
